+++ modules/spine/spine_batcher.h
@@ -84,7 +84,7 @@ public:
 
 	void add(Ref<Texture> p_texture,
 		const float* p_vertices, const float* p_uvs, int p_vertices_count,
-		const int* p_indies, int p_indies_count,
+		const unsigned short* p_indies, int p_indies_count,
 		Color *p_color, bool flip_x, bool flip_y);
 
 	void add_set_blender_mode(int p_mode);
